script to pull info from YouTube API into a txt file
This script uses Node to make a call to the YouTube API to get video
information for all videos on a given channel. Once the call is made,
the JSON gets parsed, and the script makes as many calls as it needs to
in order to get through all videos on the channel. The API returns 50
results max per call so it will make one call for every 50 videos on the
channel you're wanting data from. You end up with a text file that can
easily be opened in Excel as well.